task_manager/requirements.txt
Django==3.1.1
djangorestframework==3.12.1
psycopg2==2.8.6
pytz==2020.1
sqlparse==0.3.1
djoser==2.0.5
django-simple-history==2.11.0
